CIT-9-SA

+++ src/main/java/Citronix/dto/records/field/FieldResponseDTO.java
@@ -9,7 +9,7 @@ public record FieldResponseDTO(
         @NotNull
         UUID id,
         @NotNull
-        long superficie
+        double superficie
 ) {
 
 }
